requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.11.0
asgiref==3.10.0
audioread==3.1.0
av==12.3.0
babel==2.17.0
boto3==1.40.67
botocore==1.40.67
brotli==1.2.0
cached_path==1.8.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.4
click==8.3.0
cn2an==0.5.22
coloredlogs==15.0.1
ctranslate2==4.6.0
Cython==3.2.0
dateparser==1.1.8
decorator==5.2.1
Deprecated==1.3.1
Distance==0.1.3
Django==5.2.7
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
eng_to_ipa==0.0.2
fastapi==0.121.0
faster-whisper==1.0.3
ffmpy==0.6.4
filelock==3.19.1
flatbuffers==25.9.23
fsspec==2025.9.0
fugashi==1.3.0
g2p-en==2.1.0
g2pkk==0.1.2
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
grpcio==1.76.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
inflect==7.0.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonlines==1.2.0
langid==1.1.6
lazy_loader==0.4
librosa==0.9.1
llvmlite==0.45.1
loguru==0.7.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mecab-python3==1.0.9
melotts @ git+https://github.com/myshell-ai/MeloTTS.git@209145371cff8fc3bd60d7be902ea69cbdb7965a
mpmath==1.3.0
msgpack==1.1.2
MyShell-OpenVoice @ git+https://github.com/myshell-ai/OpenVoice.git@74a1d147b17a8c3092dd5430504bd83ef6c7eb23
networkx==2.8.8
nltk==3.9.2
num2words==0.5.12
numba==0.62.1
numpy==1.26.4
onnxruntime==1.23.2
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plac==1.4.5
platformdirs==4.5.0
pooch==1.8.2
proces==0.1.7
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pykakasi==2.2.1
pypinyin==0.50.0
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
regex==2025.11.3
requests==2.32.5
resampy==0.4.3
rich==13.9.4
rsa==4.9.1
ruff==0.14.3
s3transfer==0.14.0
safehttpx==0.1.7
scikit-learn==1.7.2
scipy==1.16.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
sqlparse==0.5.3
starlette==0.49.3
sympy==1.14.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.13.3
tomlkit==0.13.3
torch==2.9.0
torchaudio==2.9.0
tqdm==4.67.1
transformers==4.27.4
txtsplit==1.0.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.3.7
unidic==1.1.0
unidic-lite==1.0.8
urllib3==2.5.0
uvicorn==0.38.0
wasabi==0.10.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==2.0.0
